renamed file, added test description, replaced modules to be enable, dropped unrelated tests
